Changed packaging around and added some lines to the test program
